+++ video_thumbnail/lib/video_thumbnail.dart
@@ -22,7 +22,7 @@ class VideoThumbnail {
   /// If the thumbnailPath is ommited or null, a thumbnail image file will be created under the same folder as the video file.
   /// Specify the maximum height or width for the thumbnail or 0 for same resolution as the original video.
   /// The lower quality value creates lower quality of the thumbnail image, but it gets ignored for PNG format.
-  static Future<XFile?> thumbnailFile({
+  static Future<XFile> thumbnailFile({
     required String video,
     Map<String, String>? headers,
     String? thumbnailPath,
@@ -33,9 +33,6 @@ class VideoThumbnail {
     int quality = 10,
   }) async {
     assert(video.isNotEmpty);
-    if (video.isEmpty) {
-      return null;
-    }
 
     return VideoThumbnailPlatform.instance.thumbnailFile(
       video: video,
@@ -53,7 +50,7 @@ class VideoThumbnail {
   /// The video can be a local video file, or an URL repreents iOS or Android native supported video format.
   /// Specify the maximum height or width for the thumbnail or 0 for same resolution as the original video.
   /// The lower quality value creates lower quality of the thumbnail image, but it gets ignored for PNG format.
-  static Future<Uint8List?> thumbnailData({
+  static Future<Uint8List> thumbnailData({
     required String video,
     Map<String, String>? headers,
     ImageFormat imageFormat = ImageFormat.PNG,