+++ video_thumbnail/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.6.0
 * Thanks for maRci002
-  - **Breaking change**: `VideoThumbnail.thumbnailFile` now returns `XFile?` instead of `String?`
+  - **Breaking change**: `VideoThumbnail.thumbnailFile` now returns `Future<XFile>` instead of `Future<String?>`
+  - **Breaking change**: `VideoThumbnail.thumbnailData` now returns `Future<Uint8List>` instead of `Future<Uint8List?>`
   - migrating to use platform interface
   - add web implementation